2024.02.29 디렉토리 명  변경, 유레카 클라이언트 설정 추가
1. 중간 경로명 프로젝트명으로 수정
2. 유레카 클라이언트 디펜던시 추가
3. 유레카 클라이언트 설정 추가
