--- ahub-fair-and-news/src/main/java/com/teampheonix/fairnews/AhubFairAndNewsApplication.java
@@ -1,13 +1,0 @@
-package com.teampheonix.fairnews;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AhubFairAndNewsApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(AhubFairAndNewsApplication.class, args);
-	}
-
-}
