+++ ahub-fair-and-news/src/main/java/com/teamphoenix/ahub/fair/query/configuration/FairConfiguration.java
@@ -1,0 +1,10 @@
+package com.teamphoenix.ahub.fair.query.configuration;
+
+import org.apache.ibatis.annotations.Mapper;
+import org.mybatis.spring.annotation.MapperScan;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@MapperScan(basePackages = "com.teamphoenix.ahub.fair.query.mapper", annotationClass = Mapper.class )
+public class FairConfiguration {
+}
