--- ahub-fair-and-news/src/main/java/com/teamphoenix/fairnews/news/query/configuration/NewsConfiguration.java
@@ -1,10 +1,0 @@
-package com.teamphoenix.fairnews.news.query.configuration;
-
-import org.apache.ibatis.annotations.Mapper;
-import org.mybatis.spring.annotation.MapperScan;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@MapperScan(basePackages = "com.teamphoenix.fairnews.news.query.mapper", annotationClass = Mapper.class)
-public class NewsConfiguration {
-}
