--- ahub-fair-and-news/src/main/java/com/teamphoenix/fairnews/news/query/mapper/NewsMapper.java
@@ -1,10 +1,0 @@
-package com.teamphoenix.fairnews.news.query.mapper;
-
-import com.teamphoenix.fairnews.news.query.dto.NewsDTO;
-import org.apache.ibatis.annotations.Mapper;
-
-@Mapper
-public interface NewsMapper {
-
-    NewsDTO getNewsPost(int postNum);
-}
