+++ ahub-fair-and-news/src/main/resources/application.yaml
@@ -1,0 +1,16 @@
+spring:
+    datasource:
+        driver-class-name: org.mariadb.jdbc.Driver
+        url: jdbc:mariadb://localhost:3306/ahubdb
+        username: root
+        password: root
+
+    application:
+        name: user-service
+
+eureka:
+    client:
+        register-with-eureka: true
+        fetch-registry: true
+        service-url:
+            defaultZone: http://localhost:10003/eureka
