--- ahub-fair-and-news/src/test/java/com/teampheonix/fairnews/AhubFairAndNewsApplicationTests.java
@@ -1,13 +1,0 @@
-package com.teampheonix.fairnews;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AhubFairAndNewsApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
