+++ ahub-fair-and-news/src/test/java/com/teamphoenix/ahub/AhubFairAndNewsApplicationTests.java
@@ -1,0 +1,13 @@
+package com.teamphoenix.ahub;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AhubFairAndNewsApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
